Fix issues with pickle in py3

+++ mc_luigi/taskSelection.py
@@ -109,9 +109,9 @@ def get_local_features_for_multiinp():
 
     feature_tasks = []
     for i in range(len(segs)):
-        inp0 = nInpPerSeg * i
-        inp1 = nInpPerSeg * i + 1
-        inp2 = nInpPerSeg * i + 2
+        inp0 = int(nInpPerSeg * i)
+        inp1 = int(nInpPerSeg * i + 1)
+        inp2 = int(nInpPerSeg * i + 2)
 
         feature_tasks.append([])
